change the PageStore#changeCount field from an int to a long, to cope with databases with very high transaction rates.

+++ trunk/h2/src/main/org/h2/store/Page.java
@@ -74,7 +74,7 @@ public abstract class Page extends CacheObject {
     /**
      * When this page was changed the last time.
      */
-    protected int changeCount;
+    protected long changeCount;
 
     /**
      * Copy the data to a new location, change the parent to point to the new
